refactor: move SSE event parsing into OpenAPIService, return SitemapEventMessage from protocol
openHABSitemapWidgetEvents now returns any AsyncSequence<SitemapEventMessage, any Error> & Sendable,
eliminating SitemapEventStream's dependency on OpenAPIRuntime and HTTPTypes.
The 404 → SitemapSseError.unsupported mapping also moves to OpenAPIService.
